+++ server/src/utils.js
@@ -1,0 +1,5 @@
+export const capitalize = (string) => {
+  return (string).charAt(0).toUpperCase() + (string).slice(1)
+}
+
+export default capitalize
